fitto

+++ Primes.java
@@ -16,7 +16,7 @@ public class Primes {
        }
        
        
-         int  percent = (int)(times * (double)(100 /n));
+         int  percent = (int)(times * (100.0 /n));
        
        System.out.println("There are " + times + " primes between 2 and " + n + " (" + percent + "% are primes)");
     }
